major project refactor
