+++ include/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_library( natprolib INTERFACE )
+add_subdirectory( range_queries )
