+++ include/range_queries/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library( npl_range_queries INTERFACE )
